src/requirements.txt
boto3==1.26.123
modal_client>=0.56
python-dotenv==1.0.0
fastapi
requests
dkimpy